requirements.txt
google-cloud-storage~=1.31.2
numpy
torchvision
pillow
setuptools
torch~=1.6.0
matplotlib
scikit-image